+++ app/model/pattern_matchings.js
@@ -248,6 +248,7 @@ var PatternMatchings = {
             datetime: joi.boolean(),
             url: joi.boolean(),
             showSpecies: joi.boolean(),
+            showFrequency: joi.boolean(),
         }),
         limit:  joi.number(),
         offset: joi.number(),
@@ -286,15 +287,17 @@ var PatternMatchings = {
             if (show.url) {
                 builder.addProjection('R.`recording_id` as recording_id');
             }
+            if (!show.url) {
+                builder.addProjection('PMR.`uri`');
+            }
             if (show.showSpecies) {
                 builder.addTable('JOIN species Sp ON PMR.species_id = Sp.species_id');
                 builder.addTable('JOIN songtypes St ON PMR.songtype_id = St.songtype_id');
                 builder.addProjection('Sp.scientific_name as species, St.songtype as songtype');
             }
-            if (!show.url) {
-                builder.addProjection('PMR.`uri`');
+            if (show.showFrequency) {
+                builder.addProjection('R.sample_rate as sample_rate');
             }
-
             if(!show.names){
                 builder.addProjection(
                     'PMR.`recording_id`',
@@ -615,6 +618,10 @@ var PatternMatchings = {
             pmr.url = `${config('hosts').publicUrl}/api/project/${projectUrl}/recordings/download/${pmr.recording_id}`;
             delete pmr.recording_id;
         }
+        if (pmr.sample_rate) {
+            pmr.frequency = pmr.sample_rate / 2;
+            delete pmr.sample_rate;
+        }
         pmr.meta = pmr.meta ? PatternMatchings.__parse_meta_data(pmr.meta) : null;
         const namePartials = pmr.recording.split('/');
         pmr.recording = pmr.meta && pmr.meta.filename? pmr.meta.filename : namePartials[namePartials.length - 1];
@@ -693,7 +700,7 @@ var PatternMatchings = {
             expertCSValidations: options.expertCSValidations,
             countCSValidations: options.countCSValidations,
             perUserCSValidations: options.perUserCSValidations,
-            show: { names: true, url: true, showSpecies: true },
+            show: { names: true, url: true, showSpecies: true, showFrequency: true },
         }).then(
             builder => dbpool.streamQuery({
                 sql: builder.getSQL(),